+++ core/src/main/java/be/appwise/core/ui/base/BaseViewModel.kt
@@ -4,17 +4,23 @@ import androidx.lifecycle.LiveData
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
+import be.appwise.core.util.SingleLiveEvent
 import kotlinx.coroutines.CoroutineExceptionHandler
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.plus
 
-abstract class BaseViewModel : ViewModel() {
+open class BaseViewModel : ViewModel() {
     @Suppress("MemberVisibilityCanBePrivate")
     var vmScope = viewModelScope
 
+    private val _coroutineException = SingleLiveEvent<Throwable?>()
+    val coroutineException: LiveData<Throwable?> = _coroutineException
+    fun setCoroutineException(exception: Throwable?) {
+        _coroutineException.value = exception
+    }
+
     private val _loading = MutableLiveData<Boolean>().apply { value = false }
     val loading get() = _loading as LiveData<Boolean>
-
     fun isLoading(loading: Boolean) {
         _loading.postValue(loading)
     }
@@ -29,14 +35,30 @@ abstract class BaseViewModel : ViewModel() {
         showLoading(onSuccess)
     }
 
-    fun setDefaultExceptionHandler(onError: (error: Throwable) -> Unit = {}) {
+    private fun setDefaultExceptionHandler(onError: (error: Throwable) -> Unit = {}) {
         vmScope = vmScopeWithCustomExceptionHandler(onError)
     }
 
-    @Suppress("MemberVisibilityCanBePrivate")
-    fun vmScopeWithCustomExceptionHandler(onError: (error: Throwable) -> Unit = {}) =
+    /**
+     * Function that can be used to have a custom exception handler per coroutine launch
+     *
+     * ```
+     * fun addToSomething(onFinish: (Throwable?) -> Unit) = vmScopeWithCustomExceptionHandler { onFinish(it) }.launch {
+     *     callSomething {}
+     *     onFinish(null)
+     * }
+     * ``
+     *
+     * Can also be overridden in a viewModel to make it more reusable
+     */
+    open fun vmScopeWithCustomExceptionHandler(onError: (error: Throwable) -> Unit = {}) =
         (viewModelScope + CoroutineExceptionHandler { _, throwable ->
             isLoading(false)
             onError(throwable)
+            setCoroutineException(throwable)
         })
+
+    init {
+        setDefaultExceptionHandler()
+    }
 }